requirements.txt
crewai
crewai[tools]
langchain-google-genai
langchain-community
langchain-tavily
tavily-python
python-dotenv